refactor: modularize input management
